requirements.txt
pyspark
matplotlib
apache-airflow-providers-apache-spark